Addressed various PR comments
