+++ api/src/services/attachment-service.ts
@@ -20,7 +20,8 @@ import {
   SurveyAttachmentPublish,
   SurveyReportPublish
 } from '../repositories/history-publish-repository';
-import { deleteFileFromS3, generateS3FileKey } from '../utils/file-utils';
+import { deleteFileFromS3, generateS3FileKey, getFileFromS3 } from '../utils/file-utils';
+import { BctwService, IBctwUser } from './bctw-service';
 import { DBService } from './db-service';
 import { HistoryPublishService } from './history-publish-service';
 import { PlatformService } from './platform-service';
@@ -973,5 +974,37 @@ export class AttachmentService extends DBService {
 
     // Delete the attachment from S3
     await deleteFileFromS3(attachment.key);
+  }
+
+  /**
+   * Inserts a keyx reference for the given attachment ID.
+   *
+   * @param {number} attachmentId
+   * @return {*}
+   * @memberof AttachmentKeyxService
+   */
+  async insertKeyxReference(attachmentId: number) {
+    return this.attachmentRepository.insertSurveyAttachmentKeyx(attachmentId);
+  }
+
+  /**
+   * Processes keyx records.
+   * TODO: This can likely be made much more efficient.
+   *
+   * @param {IBctwUser} user
+   * @return {*}  {Promise<void>}
+   * @memberof AttachmentKeyxService
+   */
+  async processKeyxRecords(user: IBctwUser): Promise<void> {
+    const bctwService = new BctwService(user);
+
+    // get list of unprocessed records
+    const records = await this.attachmentRepository.getSurveyKeyxAttachmentsToProcess();
+    // get files from s3
+    for (const record of records) {
+      const file = await getFileFromS3(record.key);
+      await bctwService.uploadKeyX(file, record.key);
+      await this.attachmentRepository.updateSurveyAttachmentKeyx(record.survey_attachment_id);
+    }
   }
 }
